# api/openapi.yaml
openapi: 3.0.3
info:
  title: TalkFlow API
  description: API for TalkFlow, a smart video/audio call application with AI-powered notes.
  version: 1.0.0
  contact:
    name: TalkFlow Support
    email: support@talkflow.example.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.talkflow.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
components:
  schemas:
...
    Meeting:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        hostId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MeetingStatus'
        notesEnabled:
          type: boolean
        roomCode:
          type: string
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Participant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/ParticipantRole'
        joinedAt:
          type: string
          format: date-time
          nullable: true
        leftAt:
          type: string
          format: date-time
          nullable: true
...
    MeetingStatus:
      type: string
      enum: [SCHEDULED, ACTIVE, ENDED, CANCELLED]
    ParticipantRole:
      type: string
      enum: [HOST, PARTICIPANT]
...
    MeetingInvitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/InvitationStatus'
        invitedAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
    InvitationStatus:
      type: string
      enum: [PENDING, ACCEPTED, DECLINED, EXPIRED]
paths:
...
  /meetings:
    get:
      tags:
        - Meetings
      summary: Get a list of meetings for the current user
      responses:
        '200':
          description: A list of meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meeting'
    post:
      tags:
        - Meetings
      summary: Create a new meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                notesEnabled:
                  type: boolean
      responses:
        '201':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
  /meetings/{meetingId}:
    get:
      tags:
        - Meetings
      summary: Get a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested meeting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
    patch:
      tags:
        - Meetings
      summary: Update a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                notesEnabled:
                  type: boolean
      responses:
        '200':
          description: Meeting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
    delete:
      tags:
        - Meetings
      summary: Delete a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Meeting deleted successfully
  /meetings/{meetingId}/join:
    post:
      tags:
        - Meetings
      summary: Join a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Joined meeting successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
  /meetings/{meetingId}/leave:
    post:
      tags:
        - Meetings
      summary: Leave a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Left meeting successfully
  /meetings/{meetingId}/start:
    post:
      tags:
        - Meetings
      summary: Start a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
  /meetings/{meetingId}/end:
    post:
      tags:
        - Meetings
      summary: End a specific meeting
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Meeting ended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
...
  /meetings/{meetingId}/invite:
    post:
      tags:
        - Meetings
      summary: Invite users to a meeting via email or phone
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                invitations:
                  type: array
                  items:
                    type: object
                    properties:
                      email:
                        type: string
                        format: email
                        nullable: true
                      phone:
                        type: string
                        nullable: true
                    anyOf:
                      - required: [email]
                      - required: [phone]
      responses:
        '201':
          description: Invitations sent successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MeetingInvitation'
  /meetings/join/{roomCode}:
    post:
      tags:
        - Meetings
      summary: Join a meeting using room code
      parameters:
        - name: roomCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Joined meeting successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meeting:
                    $ref: '#/components/schemas/Meeting'
                  participant:
                    $ref: '#/components/schemas/Participant'
...